STABLE14-macos104-no-64bit-resid-yet-20060825
we're not ready for 64bit resid yet (in dir code)


(cherry picked from commit 3cdf27d6f070dd93c48995c344fba3b9ab0ac48e)

+++ src/afs/afs_osi.h
@@ -398,7 +398,7 @@ typedef struct timeval osi_timeval_t;
 
 #ifdef AFS_DARWIN80_ENV
 #define AFS_UIO_OFFSET(uio) uio_offset(uio)
-#define AFS_UIO_RESID(uio) uio_resid(uio)
+#define AFS_UIO_RESID(uio) (int)uio_resid(uio)
 #define AFS_UIO_SETOFFSET(uio, off) uio_setoffset(uio, off)
 #define AFS_UIO_SETRESID(uio, val) uio_setresid(uio, val)
 #else
